gltfpack: Remove support for command line texture compression
To simplify gltfpack maintenance, we no longer support using toktx or
basisu as command line executables. Instead, we use the builtin BasisU
library if gltfpack was built with it, which allows us to use optimally
threaded compression, or simply error.

For node.js builds we link the native builds instead.