simplify: Fix classification around degenerate triangles
When a triangle is topologically degenerate, it has three edges, two of them
degenerate and one normal. The normal edge can close an adjacent open edge.
Normally this is fine because the extra two edges extend the boundary for
topological analysis, but when the extra two edges are degenerate, they are
considered closed by definition, as such degenerate triangles around borders
or seams may result in classifying some of the vertices as manifold.

This can be solved by pruning degenerate triangles during updateEdgeAdjacency,
but for now this change suggests a simpler fix that effectively locks the
degenerate vertex.

Fixes #312.